added auth checks for admin and team captain

+++ server/routes/league.router.js
@@ -57,7 +57,16 @@ router.post('/join', rejectUnauthenticated, (req, res) => {
     });
 });
 
-router.post('/', (req, res) => {
+router.post('/', rejectUnauthenticated, (req, res) => {
+
+  // Only admins can add a league
+  if (req.user.authLevel !== 'ADMIN') {
+    res.sendStatus(403);
+
+    // IMPORTANT! Stop the rest of the function from running
+    return;
+  }
+
   const newLeague = req.body;
   console.log('new league', newLeague);
   let queryText = `
@@ -80,8 +89,15 @@ router.post('/', (req, res) => {
     });
 });
 
-router.put('/saveEdits', (req, res) => {
-  // console.log('req.body saveEdits in League.router', req.body);
+router.put('/saveEdits', rejectUnauthenticated, (req, res) => {
+
+  // Only admins can edit leagues
+  if (req.user.authLevel !== 'ADMIN') {
+    res.sendStatus(403);
+
+    // IMPORTANT! Stop the rest of the function from running
+    return;
+  }
 
   let queryText = `
     UPDATE "leagues"
@@ -101,7 +117,14 @@ router.put('/saveEdits', (req, res) => {
 })
 
 router.delete('/delete/:id', rejectUnauthenticated, (req, res) => {
-  // console.log('req.params deleteLeague in League.router', deleteLeague );
+  
+  // Only admins can delete a league
+  if (req.user.authLevel !== 'ADMIN') {
+    res.sendStatus(403);
+
+    // IMPORTANT! Stop the rest of the function from running
+    return;
+  }
 
   const deleteQuery = `DELETE FROM "leagues" WHERE "id" = $1;`;
